ig-reels-research/requirements.txt
playwright>=1.40.0
Pillow>=9.0.0
opencv-python>=4.8.0
imagehash>=4.3.0
pytesseract>=0.3.10
